pasteHandler in editor-settings
